product_service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
boolean.py==4.0
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.4
cryptography==43.0.3
cyclonedx-python-lib==7.6.2
dateparser==1.2.0
defusedxml==0.7.1
Django==5.1.2
django-allauth==65.1.0
django-axes==7.0.0
django-cors-headers==4.6.0
django-defender==0.9.8
django-fernet-fields==0.6
django-js-asset==2.2.0
django-mptt==0.16.0
django-otp==1.5.4
django-ratelimit==4.1.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
html5lib==1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
license-expression==30.4.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packageurl-python==0.16.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
psycopg2-binary==2.9.10
py-serializable==1.1.2
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
simplejson==3.19.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
toml==0.10.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
websockets==14.1
yarl==1.17.1
zeep==4.3.1
